Problem correction
Corrected problems found in HW2 and started using linter
