+++ cmd/server/main.go
@@ -1,19 +1,12 @@
-package server
+package main
 
 import (
+	"log"
 	"net/http"
 
-	"github.com/nndergunov/RTGC-Project/api"
-	"github.com/nndergunov/RTGC-Project/internal/httpserver"
+	"github.com/nndergunov/RTGC-Project/cmd/server/config"
 )
 
-func Main() {
-	http.ListenAndServe(":8080", New())
-}
-
-func New() http.Handler {
-	mux := httpserver.MainServer()
-	a := &api.API{Mux: mux}
-	a.Init()
-	return a
+func main() {
+	log.Fatal(http.ListenAndServe(":8080", config.New()))
 }
