--- internal/chat/message.go
@@ -1,12 +1,0 @@
-package chat
-
-import (
-	"time"
-)
-
-type Message struct {
-	FromID string
-	ToID   string
-	Text   string
-	Time   time.Time
-}
