--- main.go
@@ -1,12 +1,0 @@
-package main
-
-import (
-	"fmt"
-
-	"github.com/nndergunov/RTGC-Project/cmd/server"
-)
-
-func main() {
-	fmt.Printf("\nmain func started\n")
-	server.Main()
-}
