+++ pkg/domain/message.go
@@ -1,0 +1,12 @@
+package domain
+
+import (
+	"time"
+)
+
+type Message struct {
+	FromUserID string
+	ToGroupID  string
+	Text       string
+	Time       time.Time
+}
