+++ pkg/domain/user.go
@@ -1,4 +1,4 @@
-package v1
+package domain
 
 type UserCreateRequest struct {
 	ID    string   `json:"id"`
